correction on display errors message (upload)

+++ controllers/user.controller.js
@@ -21,19 +21,22 @@ module.exports.updateUser = async (req, res) => {
         return res.status(400).send('ID unknown : ' + req.params.id)
 
     try {
-        UserModel.findOneAndUpdate(        //la méthode findOneAndUpdate prend en 1er paramétre un filtre pour trouver le document à update et en 2eme l'element de ce document à update
-            {_id: req.params.id},
-            {
-                $set: {              //l'opérateur '$set' remplace la valeur d'un champ avec la valeur spécifié
-                    bio: req.body.bio
+        if (req.auth.userId != req.params.id) {
+            return res.status(401).send('Unauthorized User');
+        } else {
+            UserModel.findOneAndUpdate(        //la méthode findOneAndUpdate prend en 1er paramétre un filtre pour trouver le document à update et en 2eme l'element de ce document à update
+                {_id: req.params.id},
+                {
+                    $set: {              //l'opérateur '$set' remplace la valeur d'un champ avec la valeur spécifié
+                        bio: req.body.bio
+                    }
+                },
+                { new: true, upsert: true, setDefaultOnInsert: true}, //option new : retourne le document après l'update, upsert : si aucun document ne correspond au filtre mongo le crée, setDefaultOnInsert : si il y a upsert mongo insérera valeur par défaut
+                (err, data) => {
+                    if (!err) return res.send(data);
+                    if (err) return res.status(400).json({ message: err});
                 }
-            },
-            { new: true, upsert: true, setDefaultOnInsert: true}, //option new : retourne le document après l'update, upsert : si aucun document ne correspond au filtre mongo le crée, setDefaultOnInsert : si il y a upsert mongo insérera valeur par défaut
-            (err, data) => {
-                if (!err) return res.send(data);
-                if (err) return res.status(400).json({ message: err});
-            }
-        )
+            )}
     } catch (err) {
         return res.status(500).json({ message: err});
     }
@@ -44,8 +47,12 @@ module.exports.deleteUser = async (req, res) => {
         return res.status(400).send('ID unknown : ' + req.params.id)
     
         try {
-            await UserModel.deleteOne({ _id: req.params.id }).exec();
-            res.status(200).json({ message: "User deleted. "});
+            if (req.auth.userId != req.params.id) {
+                return res.status(401).send('Unauthorized User');
+            } else {
+                await UserModel.deleteOne({ _id: req.params.id }).exec();
+                res.status(200).json({ message: "User deleted. "});
+            }
         } catch (err) {
             return res.status(500).json({ message: err});
         }
